DXP-1633 Reduce number of indexer classes layers and improve class naming
