+++ src/catalog/Indexer/CategoriesIndexer.php
@@ -1,35 +1,36 @@
 <?php declare(strict_types=1);
 
-namespace StreamX\ConnectorCatalog\Model\Indexer;
+namespace StreamX\ConnectorCatalog\Indexer;
 
 use Psr\Log\LoggerInterface;
-use StreamX\ConnectorCatalog\Indexer\CategoryIndexerHandler;
-use StreamX\ConnectorCatalog\Model\Indexer\Action\Category as CategoryAction;
+use StreamX\ConnectorCatalog\Model\Indexer\Action\CategoryAction;
+use StreamX\ConnectorCatalog\Model\Indexer\CategoryProcessor;
+use StreamX\ConnectorCore\Api\IndexersConfigInterface;
 use StreamX\ConnectorCore\Config\OptimizationSettings;
+use StreamX\ConnectorCore\Indexer\BaseStreamxIndexer;
 use StreamX\ConnectorCore\Indexer\IndexableStoresProvider;
 use StreamX\ConnectorCore\Streamx\ClientResolver;
 use StreamX\ConnectorCore\System\GeneralConfig;
 
-class Category extends BaseStreamxIndexer {
-
+class CategoriesIndexer extends BaseStreamxIndexer
+{
     public function __construct(
         GeneralConfig $connectorConfig,
-        CategoryIndexerHandler $indexerHandler,
         IndexableStoresProvider $indexableStoresProvider,
-        CategoryAction $action,
+        CategoryAction $categoryAction,
         LoggerInterface $logger,
         OptimizationSettings $optimizationSettings,
-        ClientResolver $clientResolver
+        ClientResolver $clientResolver,
+        IndexersConfigInterface $indexersConfig
     ) {
         parent::__construct(
             $connectorConfig,
-            $indexerHandler,
             $indexableStoresProvider,
-            $action,
+            $categoryAction,
             $logger,
             $optimizationSettings,
             $clientResolver,
-            'Categories'
+            $indexersConfig->getByName(CategoryProcessor::INDEXER_ID)
         );
     }
 }
