+++ src/catalog/etc/indexer.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Indexer/etc/indexer.xsd">
     <indexer id="streamx_product_indexer" view_id="streamx_product_indexer"
-             class="StreamX\ConnectorCatalog\Model\Indexer\Product">
+             class="StreamX\ConnectorCatalog\Indexer\ProductsIndexer">
         <title translate="true">StreamX Product Indexer</title>
         <description translate="true">Updates products in StreamX</description>
         <dependencies>
@@ -10,12 +10,12 @@
         </dependencies>
     </indexer>
     <indexer id="streamx_category_indexer" view_id="streamx_category_indexer"
-             class="StreamX\ConnectorCatalog\Model\Indexer\Category">
+             class="StreamX\ConnectorCatalog\Indexer\CategoriesIndexer">
         <title translate="true">StreamX Category Indexer</title>
         <description translate="true">Updates categories in StreamX</description>
     </indexer>
     <indexer id="streamx_attribute_indexer" view_id="streamx_attribute_indexer"
-             class="StreamX\ConnectorCatalog\Model\Indexer\Attribute">
+             class="StreamX\ConnectorCatalog\Indexer\AttributesIndexer">
         <title translate="true">StreamX Attribute Indexer</title>
         <description translate="true">Updates products with added or modified attributes in StreamX</description>
     </indexer>
