+++ src/catalog/etc/mview.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Mview/etc/mview.xsd">
-    <view id="streamx_product_indexer" class="StreamX\ConnectorCatalog\Model\Indexer\Product" group="indexer">
+    <view id="streamx_product_indexer" class="StreamX\ConnectorCatalog\Indexer\ProductsIndexer" group="indexer">
         <subscriptions>
             <!-- TODO verify on all magento versions. Some of them may not have the entity_id column in the attributes tables -->
             <table name="catalog_product_entity_datetime" entity_column="entity_id" />
@@ -18,7 +18,7 @@
         </subscriptions>
     </view>
 
-    <view id="streamx_category_indexer" class="StreamX\ConnectorCatalog\Model\Indexer\Category" group="indexer">
+    <view id="streamx_category_indexer" class="StreamX\ConnectorCatalog\Indexer\CategoriesIndexer" group="indexer">
         <subscriptions>
             <table name="catalog_category_entity" entity_column="entity_id" />
             <table name="catalog_category_entity_decimal" entity_column="entity_id" />
@@ -29,7 +29,7 @@
         </subscriptions>
     </view>
 
-    <view id="streamx_attribute_indexer" class="StreamX\ConnectorCatalog\Model\Indexer\Attribute" group="indexer">
+    <view id="streamx_attribute_indexer" class="StreamX\ConnectorCatalog\Indexer\AttributesIndexer" group="indexer">
         <subscriptions>
             <table name="eav_attribute" entity_column="attribute_id" />
             <table name="catalog_eav_attribute" entity_column="attribute_id" />
